approval-threshold: lower it daily after three days with no commits
Right now, if someone disappears then we'll have a draw; see #19.  This change is a straight-forward way around this: once three days have passed without a commit, then for every additional day we require one fewer approver.

I'm open to other functions, but I'd like to get something in soon because we're adding a lot of players.

(I think this is probably not what we'll want long term.)