players: switch to a directory
Currently, when new players are added that makes merge conflicts, and requires
a lot of re-approvals.  An easy way to avoid this is to use a directory instead
of a file.

Also add everyone who has asked to be added:
 * Closes #11
 * Closes #12
 * Closes #15
 * Closes #20
 * Closes #22

We should probably merge #23 first, so we don't get stuck if we lose someone
(#19).