approval-threshold: allow changes to go in with majority approval if it's been sitting for a day with no rejections
At some point we may want to allow things to go in over someone's rejection: this is not that.  Any player can still reject any PR and stop it.  But if a PR has no rejections, has been approved by a majority of players, and not been updated in the last day, allow it to go in.

Also refactor time-related functions.