Add get_user_points
Add a function to pull points out of the points files, and print them.  Still
doesn't make points do anything.  Pulled out of #33 to make things easier to
review.

During review of #33 people noted that you might be able to make this function
fail with an Exception saying that you win, which might not be caught earlier
because get_user_points would only be called in determine_if_winner, which
isn't called on PR validation.  One way to handle this would be to rewrite the
code defensively, catching every potential issue (something not being a
directory, contents of the file not being an int, etc), but this PR takes a
much simpler (and easier to trust approach) of just additionally running
get_user_points on PR validation.  This way, if there's something wrong with
the environment the PR will fail in validation and not be mergeable.

Depends on #47, compatible with #33