named-bonuses: alternative implementation
This is an alternative to #47, where the point value is in the filename instead of the file body.  Suggested in the comments on #50, the idea is this makes it easier to parse diffs that work with points.

I also added two example files to my directory, one showing point gain and one showing point loss, so you can see what negative numbers look like.

I don't have strong feelings between this and #47, and would be happy with either.