random: don't bias towards earlier users
Currently our random selection has a bias towards earlier users.  For each user we generate a random number and see if their total is below it, so there's an edge to being earlier.  This edge gets larger as points grow.  For example, if we had two players each with 50k points, the first player would win 50% of the time, while the second would only win 50% of remainder, or 25% of the time.

Switch to picking a single random number and seeing which, if any, player that corresponds to.

Testing this with the points scalar at 0.001 instead of 0.00001 to not take so long, I see:

   8 Exception: pavellishin wins!
  10 Exception: csvoss wins!
  13 Exception: jeffkaufman wins!
  14 Exception: tnelling wins!
 955 The game continues.

which seems reasonable given our current points.