start using reproducible random numbers
Currently we have Travis running random.random() to determine if someone has won.  This of course gives different values each time you run it.  This is basically fine, but it would be better to use a hash of something that (a) we all can verify and (b) we can't predict.  Then we don't have to trust Travis or worry about people re-running Travis builds.

This PR switches us to using the hash of the merge commit as our source of randomness.  We only get one of these, so this depends on the fix in #97.