Allow old approvals to persist when it's safe to do so.
Consider #111.  It was approved at 250d3a69 but then I needed to merge from
master to get up to date.  The delta from 250d3a69 to master is identical to
the diff from the new HEAD, 98b5a3a4, to master.  Both are exactly:

```
diff --git a/validate.py b/validate.py
index 4bdd054..4debe90 100644
--- a/validate.py
+++ b/validate.py
@@ -1,6 +1,7 @@
 import os
 import runpy
 import copy
+import traceback

 import util
 import pull_request
@@ -67,7 +68,7 @@ def determine_if_mergeable(pr):

   # Go through rules sorted by priority, with ties broken by the filename.
   for rule_priority, rule_full_fname, rule_name, is_allow in sorted(rules):
-    print('Running rule %s' % rule_full_fname)
+    print('\nRunning rule %s' % rule_full_fname)

     pr_copy = copy.deepcopy(pr)

@@ -82,6 +83,7 @@ def determine_if_mergeable(pr):
           print('\nPASS: %s' % rule_name)
           return
       except Exception as e:
+        traceback.print_exc()
         print('  %s: %s' % (rule_full_fname, e))
     else:
       # Raises an exception to indicate blocking, anything else for no
```

Since the diff didn't change with merging in master, it's still safe to count
these approvals at the earlier commit as valid.  They represent approving the
same change.

If on master any of the files involved had been added, removed, or edited then
the diff would not be exactly the same and we'd need to get new approvals.

Fixes #95